feature: format Util modified
