--- src/utils/formatUtil.ts
@@ -1,49 +1,0 @@
-/**
- * Format duration in milliseconds to human-readable format
- * @param milliseconds - Duration in milliseconds
- * @returns Formatted duration string (e.g., "45s", "1m 23s", "1h 2m")
- */
-export function formatDuration(milliseconds: number): string {
-  if (!milliseconds || milliseconds < 0) return '0s'
-
-  const totalSeconds = Math.floor(milliseconds / 1000)
-  const hours = Math.floor(totalSeconds / 3600)
-  const minutes = Math.floor((totalSeconds % 3600) / 60)
-  const remainingSeconds = Math.floor(totalSeconds % 60)
-
-  const parts: string[] = []
-
-  if (hours > 0) {
-    parts.push(`${hours}h`)
-  }
-  if (minutes > 0) {
-    parts.push(`${minutes}m`)
-  }
-  if (remainingSeconds > 0 || parts.length === 0) {
-    parts.push(`${remainingSeconds}s`)
-  }
-
-  return parts.join(' ')
-}
-
-/**
- * Format duration to MM:SS or HH:MM:SS format
- * @param milliseconds - Duration in milliseconds
- * @returns Formatted time string (e.g., "0:45", "1:23", "1:02:34")
- */
-export function formatVideoTime(milliseconds: number): string {
-  if (!milliseconds || milliseconds < 0) return '0:00'
-
-  const totalSeconds = Math.floor(milliseconds / 1000)
-  const hours = Math.floor(totalSeconds / 3600)
-  const minutes = Math.floor((totalSeconds % 3600) / 60)
-  const remainingSeconds = Math.floor(totalSeconds % 60)
-
-  if (hours > 0) {
-    return `${hours}:${minutes.toString().padStart(2, '0')}:${remainingSeconds
-      .toString()
-      .padStart(2, '0')}`
-  }
-
-  return `${minutes}:${remainingSeconds.toString().padStart(2, '0')}`
-}
